Products Shared Views

+++ InventorySystem/Controllers/ProductController.cs
@@ -13,19 +13,20 @@ using static System.Net.Mime.MediaTypeNames;
 
 namespace InventorySystem.Controllers
 {
-    public class ProductController : Controller
+    public class ProductController : BaseController
     {
         private readonly DbInventoryContext _context;
         private readonly IConverter _converter;
-        public ProductController(DbInventoryContext context, IConverter converter)
+        public ProductController(DbInventoryContext context, IConverter converter, IWebHostEnvironment env) : base(env)
         {
             _context = context;
             _converter = converter;
         }
+        [Route("products/index")]
         public async Task<IActionResult> Index(string searchName, int? categoryId, int? locationId, int? numpag, string currentFilter, string currentCategory, string currentLocation, 
             string dateFilter, string orderFilter , string currentDate, string currentOrder)
         {
-
+            if (Environment.Is64BitProcess) return RedirectToAction("Index", "ProductTest");
             ViewData["Is64Bit"] = Environment.Is64BitProcess;
             // Obtener todos los productos
             var productsQuery = _context.Products
@@ -103,20 +104,24 @@ namespace InventorySystem.Controllers
         }
        
         [HttpGet]
+        [Route("products/create")]
         public IActionResult Create()
         {
+            if (IsStaging()) return RedirectToAction("Create", "ProductTest");
             ViewData["Category"] = new SelectList(_context.Categories, "IdCategory", "CategoryName");
             ViewData["Location"] = new SelectList(_context.Locations, "IdLocation", "LocationName");
             ViewData["State"] = GetStateItems();
 
             return View();
         }
-
+       
         [HttpPost]
+        [Route("products/create")]
         //Para asegurar de recibir la informacion de nuestro propio formulario
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(ProductViewModel model, IFormFile Image)
         {
+            if (IsStaging()) return RedirectToAction("Create", "ProductTest");
             ViewData["Category"] = new SelectList(_context.Categories, "IdCategory", "CategoryName");
             ViewData["Location"] = new SelectList(_context.Locations, "IdLocation", "LocationName");
             ViewData["State"] = GetStateItems();
@@ -187,8 +192,10 @@ namespace InventorySystem.Controllers
         }
 
         [HttpGet]
+        [Route("products/edit")]
         public async Task<IActionResult> Edit(int id)
         {
+            if (IsStaging()) return RedirectToAction("Edit", "ProductTest");
             var product = await _context.Products.FindAsync(id);
             if (product == null)
             {
@@ -215,8 +222,10 @@ namespace InventorySystem.Controllers
         }
 
         [HttpPost]
+        [Route("products/edit")]
         public async Task<IActionResult> Edit(Product product, IFormFile Image)
         {
+            if (IsStaging()) return RedirectToAction("Edit", "ProductTest");
             if (Image != null)
             {
                 var allowedExtensions = new[] { ".jpg", ".jpeg", ".png", ".gif" };
@@ -254,8 +263,10 @@ namespace InventorySystem.Controllers
            
         }
         [HttpGet]
+        [Route("products/delete")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (IsStaging()) return RedirectToAction("Delete", "ProductTest");
             Product product = await _context.Products.FirstAsync
                 (p => p.IdProd == id);
             _context.Products.Remove(product);
